+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [v2.0.6](https://github.com/in2workspace/in2-wallet-api/releases/tag/v2.0.6)
+### Changed
+- Changed Failure case in notification endpoint implementation
+- Added Powers
+
+## [v2.0.5](https://github.com/in2workspace/in2-wallet-api/releases/tag/v2.0.5)
+### Added
+- Bitstring type credential status
 
 ## [v2.0.4](https://github.com/in2workspace/in2-wallet-api/releases/tag/v2.0.4)
 ### Added
 - Added notification endpoint implementation
-
 
 ## [v2.0.3](https://github.com/in2workspace/in2-wallet-api/releases/tag/v2.0.3)
 ### Changed